backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic[email]==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
aiosmtplib==1.1.6
python-dotenv==0.19.0
jinja2==3.0.1
psycopg2-binary==2.9.9
alembic==1.12.1
